requirements.txt
gTTS
pysrt
edge-tts
pydub 
openai-whisper
streamlit
asyncio